pyproject.toml
[tool.poetry]
name = "easy-shop"
version = "0.1.0"
description = ""
authors = ["Sergey <one6sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.7.2"
Django = "4.2.4"
djangorestframework = "3.14.0"
Pillow = "10.0.0"
pytz = "2023.3"
sqlparse = "0.4.4"
typing_extensions = "4.7.1"
setuptools = "^70.1.0"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
